test: fix llm load tests
